[#EX-2] Chat session management
We want to have an interface, where the user can start a new chat
(similar to chagpt) and see the previous chats.&#x20;

For now we should have a list of chats on the left (with placeholder
titles for now) to click on leading to a show view (only display chat id
for now) on the right. \
Also one should have a `new` button that leads to a form (empty
currently) with a create button, to create new chats.

https://bitcrowd.atlassian.net/browse/EX-2