[#EX-3] Sign in with github (#3)
* [#EX-3] Sign in with github

Add a sign in button (with a development sign in option) that is needed
to log in to the platform. The sign in should request the users email
address at a minimum and save it alongside on the user model.

https://bitcrowd.atlassian.net/browse/EX-3

* Change config structure

* Implement github auth & mock

* Rename chat to chats

* Add accounts

* Wire up

* [#EX-4] Scope chats to users (#4)

Make sure one can only see the chat’s created yourself.&#x20;

https://bitcrowd.atlassian.net/browse/EX-4