[#EX-6] Library selection (#5)
For every library and version we need to do a separate RAG ingestion. So
we need a model that can reference chunks for library and versions that
were ingested earlier.&#x20;

When starting a new chat, users should be able to select one (for now
maybe multiple later) such library/version combination they want to ask
questions about. Add a typeahead select for choosing the right lib and
version pair and display it on the chat show and in the list on the
left. \

https://bitcrowd.atlassian.net/browse/EX-6