[#EX-55] Show library progress and errors
We want to show, which libraries are currently ingesting and where there
are errors that we need to look at potentially. \
\
This would be on a page, the user is redirected to when adding a library
and it should be accessible on the menu. \
\
It should list: \

- All current ingestions (and their state)
  - List libraries currently generating chunks
    - Maybe we get it from the oban jobs and if it is errored, discarded
      show a little warning sign that we look at it currently
    - Maybe we need to check if the oban worker needs a timeout error
      for this…
  - List libraries currently embedding chunks
    - Show a state similar to the library admin
    - Should be easy, since we can just find all libraries with at least
      one chunk having nil embeddings first
- The last successful ingestions&#x20;
  - Like in the library select\

In order to make this page nice, we should add a pubsub channel where
the embedding and ingesting is reported back after each job, so that the
page can live update.&#x20;

https://bitcrowd.atlassian.net/browse/EX-55