[#EX-64] Refactor code of admin UI to work with ingestions (#59)
* [#EX-64] Refactor code of admin UI to work with…

…ingestions

We’ve introduced the concept of ingestions, so each library can now have
multiple ingestions associated, instead of only the chunks. The admin
interface doesn’t take into account the concept of ingestions yet.&#x20;

Task: Refactor the admin UI, so that we can see the individual
ingestions for each of the libraries.

Current situation:

The admin UI (“/admin”) shows a list of libraries (and corresponding
version). There is a show button for each library that takes us to the
show view.&#x20;

On the show view we can see stats about chunks for the library.

After refactoring:

The admin UI (“/admin”) shows a list of libraries (and corresponding
version). There is a show button for each library that takes us to the
show view. This part remains the same.

On the show view we want to show a list of all ingestions associated
with the library, we also want to display the state of the ingestion.\
There should be a show button for each ingestion that takes us to the
show view for the ingestion.\
On the show view for the ingestion, we want to see the exact same stats
about chunks as we previously had on the show view for the library.

https://bitcrowd.atlassian.net/browse/EX-64

* claude first draft

* With Flop.Schema for ingestions

* Add tests for topics context functions

* Mix format

* Remove get_library_stats/1

* Factorize into components, add header and breadcrumbs

* Show state next to ingestion title

* Remove unused function

* more cosmetics

* Pluralize ingestion in the route

---------

Co-authored-by: Agathe Lenclen <agathe@bitcrowd.net>