Add active flag to system prompts
This adds a new boolean field - "active" to system prompts. The field
has a unique index on it, so that there can always only be one active
system prompt.

The migration also migrates the currently activated prompt from the runtime
configuration to be set to "active". If the configuration isn't present
any more, a random prompt from the DB will be set to as the active one.