Fix test for deleting system prompts
The test was not really testing anything: it was not checking whether
the system prompt was actually gone from the page. In fact, it checked
the opposite…

The problem was, that there are already two prompts in the database. And
as they are all from the same factory, they share the same prompt…