Move drawer under navbar on small screens
This is inspired by the problem described here: https://github.com/saadeghi/daisyui/discussions/3122

We add a `top` value to the drawer, which is only applies when the
positioning of the drawer moves from sticky to fixed (see also
https://developer.mozilla.org/en-US/docs/Web/CSS/top), when it's
covering the whole screen.

With the navbar still visible, people can close the sidebar drawer
again.