[EX-111] Activate a system prompt (#99)
* Add active flag to system prompts

This adds a new boolean field - "active" to system prompts. The field
has a unique index on it, so that there can always only be one active
system prompt.

The migration also migrates the currently activated prompt from the runtime
configuration to be set to "active".

* Add field to schema and update tests

* Fix test for deleting system prompts

The test was not really testing anything: it was not checking whether
the system prompt was actually gone from the page. In fact, it checked
the opposite…

The problem was, that there are already two prompts in the database. And
as they are all from the same factory, they share the same prompt…

* Update active badge to be based on DB column

* Allow activating system prompts in the admin UI

Activating a system prompt deactivates the currently active one.

* Get system prompt from DB instead of config

We'll keep the configuration for a smoother migration path for now.

* Update seeds to add an active system prompt

Also: update migration to only activate an existing system prompt of
there is no active one yet.

* Remove seed data from tests

I felt like this makes the tests harder to reason about.

* Use bitcrowd ecto's UUID helper