Remove migration to change production model
This removes the migration to update the production model. The migration
only includes INSERT statements and a conditional to check for the
current environment. Both are anti-patterns and will make our lives
harder…

So I opted for some raw SQL juggle instead and added the new model to
the large SQL statement, more as a way to document it.