Remove env specific insert statements from migrations
The env-specific condition was added after the migrations ran on
staging/production, so it did not occur as a problem for the deployment
(yet).

I completely removed the statements, as they mess with the test database
if we skip the conditional. For the dev environment, the seeds are
already run with `mix ecto.setup`, so the state of the database should
be fine.